fix(sqlite-storage): read removed rows before the update nulls them
